debut get color texture

+++ includes/cub_struct.h
@@ -6,7 +6,7 @@
 /*   By: jdenis <jdenis@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/03/15 18:42:07 by jdenis            #+#    #+#             */
-/*   Updated: 2024/03/25 18:39:13 by jdenis           ###   ########.fr       */
+/*   Updated: 2024/03/25 20:17:00 by jdenis           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -33,8 +33,8 @@ typedef struct s_ray
 	int		side;
 	double	eu_dist;
 	double	perp_dist;
-	double	wall_x;
-	double	wall_y;
+	int	wall_x;
+	int	wall_y;
 }				t_ray;
 
 typedef struct s_player
